requirements.txt
Django==3.0.4
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
django-filter==2.3.0
django-mysql==3.7.1
